requirements.txt
dj-database-url==0.4.2
Django==1.11.17
Pillow==2.9.0
django-embed-video==1.0.0
python-decouple==3.1
mysqlclient
django-crispy-forms
django-braces==1.8.1
bootstrap-admin
djangorestframework==3.9.1